chore: rename `JSONEditorMode` to `JSONEditorRoot`
