+++ src/lib/components/modes/JSONEditorRoot.svelte
